# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services": {
      "post": {
        "operationId": "ServiceController_create",
        "summary": "Create a new service",
        "description": "Creates a new service with the provided information",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Service creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Category not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ServiceController_findAll",
        "summary": "Get all services",
        "description": "Retrieves a paginated list of services with optional filtering options",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "subcategoryIds",
            "required": false,
            "in": "query",
            "description": "Filter by subcategory IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "type": "a"
                }
              }
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter by category ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "latitude",
            "required": false,
            "in": "query",
            "description": "User latitude for location-based filtering",
            "schema": {
              "example": 37.62332465580069,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": false,
            "in": "query",
            "description": "User longitude for location-based filtering",
            "schema": {
              "example": -0.9851654443962178,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Search radius in meters for location-based filtering",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "onlyActives",
            "required": false,
            "in": "query",
            "description": "Filter only active services (true) or all services (false)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of services retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Service"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "description": "Total number of services"
                        },
                        "page": {
                          "type": "number",
                          "description": "Current page number"
                        },
                        "limit": {
                          "type": "number",
                          "description": "Items per page"
                        },
                        "totalPages": {
                          "type": "number",
                          "description": "Total number of pages"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/{id}": {
      "get": {
        "operationId": "ServiceController_findOne",
        "summary": "Get a service by ID",
        "description": "Retrieves a specific service by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ServiceController_update",
        "summary": "Update a service",
        "description": "Updates an existing service with new information",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Service update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service or category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ServiceController_remove",
        "summary": "Delete a service",
        "description": "Permanently deletes a service from the system",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Service deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/me/published": {
      "get": {
        "operationId": "ServiceController_findMyPublishedServices",
        "summary": "Get user's published services",
        "description": "Retrieves all services published by the authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User's published services retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/{id}/availability": {
      "get": {
        "operationId": "ServiceController_getServiceAvailability",
        "summary": "Get service availability",
        "description": "Retrieves the availability of a service for a specific date",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Date to check availability (YYYY-MM-DD)",
            "schema": {
              "example": "2024-01-15",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service availability retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "serviceId": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string"
                    },
                    "availableSlots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "startTime": {
                            "type": "string",
                            "example": "09:00"
                          },
                          "endTime": {
                            "type": "string",
                            "example": "10:00"
                          },
                          "isAvailable": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reviews/service/{id}": {
      "get": {
        "operationId": "ReviewController_findByServiceId",
        "summary": "Get reviews by service",
        "description": "Retrieves all reviews for a specific service",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service ID",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews for the service retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Service not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "Manexo API",
    "description": "API documentation for Manexo application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "bookings": {
            "description": "Lista de reservas asociadas al servicio",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Booking"
            }
          }
        }
      },
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Descripción detallada del servicio",
            "example": "Ofrezco servicios de plomería residencial y comercial"
          },
          "radius": {
            "type": "number",
            "description": "Radio de cobertura del servicio",
            "example": 10
          },
          "location": {
            "description": "Ubicación del servicio",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationDto"
              }
            ]
          },
          "price": {
            "type": "number",
            "description": "Precio del servicio",
            "example": 500
          },
          "subcategory": {
            "type": "number",
            "description": "ID de la subcategoría a la que pertenece el servicio",
            "example": 1
          },
          "requiresAcceptance": {
            "type": "boolean",
            "description": "Indica si el servicio requiere aceptación manual",
            "example": false
          },
          "images": {
            "description": "Lista de URLs de imágenes del servicio",
            "example": [
              "https://example.com/image1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "description",
          "radius",
          "location",
          "price",
          "subcategory"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Descripción detallada del servicio",
            "example": "Ofrezco servicios de plomería residencial y comercial"
          },
          "radius": {
            "type": "number",
            "description": "Radio de cobertura del servicio",
            "example": 10
          },
          "location": {
            "description": "Ubicación del servicio",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationDto"
              }
            ]
          },
          "price": {
            "type": "number",
            "description": "Precio del servicio",
            "example": 500
          },
          "subcategory": {
            "type": "number",
            "description": "ID de la subcategoría a la que pertenece el servicio",
            "example": 1
          },
          "requiresAcceptance": {
            "type": "boolean",
            "description": "Indica si el servicio requiere aceptación manual",
            "example": false
          },
          "images": {
            "description": "Lista de URLs de imágenes del servicio",
            "example": [
              "https://example.com/image1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}